requirements.txt
Django==2.1.7
djangorestframework==3.7.7
django-cors-headers==2.4.0
django-extensions==2.1.3
django-filter==2.0.0
django-simple-history==2.5.1
